week03

+++ week03/3.2.convert.py
@@ -5,8 +5,11 @@
 # week 3 lab 2
 
 number = float(input("Please enter an amount:"))
+
 absolutevalue = abs(number)
-string2 = "The amount in cent is : "+str(absolutevalue * 100)
+
+# the code below was returning a .0 at the end of the result until I added the str() function
+string2 = "The amount in cent is : "+ str(int(absolutevalue * 100))
 
 print (string2)
